+++ stroom-core-shared/src/main/java/stroom/dictionary/shared/DictionaryDoc.java
@@ -73,6 +73,13 @@ public class DictionaryDoc extends Doc {
         this.imports = imports;
     }
 
+    /**
+     * @return A new builder for creating a {@link DocRef} for this document's type.
+     */
+    public static DocRef.TypedBuilder buildDocRef() {
+        return DocRef.builder(DOCUMENT_TYPE);
+    }
+
     public String getDescription() {
         return description;
     }