+++ stroom-core-shared/src/main/java/stroom/xmlschema/shared/XmlSchemaDoc.java
@@ -16,6 +16,7 @@
 
 package stroom.xmlschema.shared;
 
+import stroom.docref.DocRef;
 import stroom.docstore.shared.Doc;
 import stroom.util.shared.HasData;
 
@@ -85,6 +86,13 @@ public class XmlSchemaDoc extends Doc implements HasData {
         this.data = data;
         this.deprecated = deprecated;
         this.schemaGroup = schemaGroup;
+    }
+
+    /**
+     * @return A new builder for creating a {@link DocRef} for this document's type.
+     */
+    public static DocRef.TypedBuilder buildDocRef() {
+        return DocRef.builder(DOCUMENT_TYPE);
     }
 
     public String getDescription() {